requirements.txt
Django==4.2.4
python-dotenv
django-filter
djangorestframework
markdown
